# api-service/src/main/gen/api/openapi.yaml
openapi: 3.0.3
info:
  contact:
    email: support@chokma.com
    name: Chokma Support
  description: Document extraction and processing API
  title: Chokma API
  version: 1.0.0
servers:
- description: Production
  url: https://api.chokma.com
- description: Local development
  url: http://localhost:8080
...
paths:
...
  /tenants:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Tenant'
                type: array
          description: List of tenants
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized - authentication required
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden - insufficient permissions
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal server error
      security:
      - bearerAuth: []
      summary: List tenants
      tags:
      - Tenants
      x-accepts:
      - application/json
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTenantRequest'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdResponse'
          description: Tenant created successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad request - invalid input
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized - authentication required
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden - insufficient permissions
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Tenant key already exists
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal server error
      security:
      - bearerAuth: []
      summary: Create tenant
      tags:
      - Tenants
      x-content-type: application/json
      x-accepts:
      - application/json
  /tenants/{tenantId}:
    delete:
      parameters:
      - explode: false
        in: path
        name: tenantId
        required: true
        schema:
          $ref: '#/components/schemas/UUID'
        style: simple
      responses:
        "204":
          description: Tenant deleted successfully
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized - authentication required
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden - insufficient permissions
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Resource not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal server error
      security:
      - bearerAuth: []
      summary: Delete tenant
      tags:
      - Tenants
      x-accepts:
      - application/json
...
  /jobs/async:
    post:
      description: Create job and get pre-signed upload URL
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateJobRequest'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadUrlResponse'
          description: Job created with upload URL
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad request - invalid input
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized - authentication required
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden - insufficient permissions
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal server error
      security:
      - bearerAuth: []
      summary: Submit async job
      tags:
      - Jobs
      x-content-type: application/json
      x-accepts:
      - application/json
  /jobs/sync:
    post:
      description: Upload file and process immediately
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/_jobs_sync_post_request'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadResponse'
          description: Job processed successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad request - invalid input
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized - authentication required
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden - insufficient permissions
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal server error
      security:
      - bearerAuth: []
      summary: Submit synchronous job
      tags:
      - Jobs
      x-content-type: multipart/form-data
      x-accepts:
      - application/json
  /jobs/{jobId}:
    delete:
      parameters:
      - explode: false
        in: path
        name: jobId
        required: true
        schema:
          $ref: '#/components/schemas/UUID'
        style: simple
      responses:
        "204":
          description: Job cancelled successfully
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized - authentication required
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Resource not found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Job cannot be cancelled in current state
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal server error
      security:
      - bearerAuth: []
      summary: Cancel job
      tags:
      - Jobs
      x-accepts:
      - application/json
    get:
      parameters:
      - explode: false
        in: path
        name: jobId
        required: true
        schema:
          $ref: '#/components/schemas/UUID'
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResponse'
          description: Job details retrieved
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized - authentication required
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Resource not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal server error
      security:
      - bearerAuth: []
      summary: Get job status
      tags:
      - Jobs
      x-accepts:
      - application/json
  /jobs/{jobId}/start-parse:
    post:
      description: Trigger parsing for uploaded file
      parameters:
      - explode: false
        in: path
        name: jobId
        required: true
        schema:
          $ref: '#/components/schemas/UUID'
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParseResponse'
          description: Parse started
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized - authentication required
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Resource not found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Job already parsed or in invalid state
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal server error
      security:
      - bearerAuth: []
      summary: Start parse job
      tags:
      - Jobs
      x-accepts:
      - application/json
  /jobs/{jobId}/status/realtime:
    get:
      description: Poll for current job status from Third party
      parameters:
      - explode: false
        in: path
        name: jobId
        required: true
        schema:
          $ref: '#/components/schemas/UUID'
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResponse'
          description: Realtime status retrieved
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized - authentication required
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Resource not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal server error
      security:
      - bearerAuth: []
      summary: Get realtime job status
      tags:
      - Jobs
      x-accepts:
      - application/json
...
components:
...
  schemas:
    UUID:
      example: 123e4567-e89b-12d3-a456-426614174000
      format: uuid
      pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
      type: string
...
    ErrorResponse:
      example:
        traceId: traceId
        code: code
        details:
          key: ""
        message: message
        timestamp: 2000-01-23T04:56:07.000+00:00
      properties:
        message:
          description: Human-readable error message
          type: string
        code:
          description: Machine-readable error code
          type: string
        timestamp:
          format: date-time
          type: string
        traceId:
          description: Request trace ID for debugging
          type: string
        details:
          additionalProperties: true
          description: Additional error details
          type: object
      required:
      - code
      - message
      type: object
...
    Tenant:
      example:
        createdAt: 2024-01-05T16:15:50Z
        name: Acme Corporation
        id: 123e4567-e89b-12d3-a456-426614174000
        key: acme-corp
        status: active
      properties:
        id:
          example: 123e4567-e89b-12d3-a456-426614174000
          format: uuid
          pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
          type: string
        key:
          description: Unique tenant key
          example: acme-corp
          type: string
        name:
          description: Tenant display name
          example: Acme Corporation
          type: string
        status:
          enum:
          - active
          - suspended
          - deleted
          example: active
          type: string
        createdAt:
          example: 2024-01-05T16:15:50Z
          format: date-time
          type: string
      required:
      - createdAt
      - id
      - key
      - name
      - status
      type: object
    CreateTenantRequest:
      example:
        name: name
        key: key
      properties:
        key:
          description: Unique tenant identifier
          maxLength: 64
          minLength: 3
          pattern: "^[a-z0-9-]+$"
          type: string
        name:
          description: Tenant display name
          maxLength: 128
          minLength: 1
          type: string
      required:
      - key
      - name
      type: object
...
    CreateJobRequest:
      example:
        fileName: fileName
        processorId: 123e4567-e89b-12d3-a456-426614174000
        projectId: 123e4567-e89b-12d3-a456-426614174000
        contentType: application/pdf
      properties:
        projectId:
          example: 123e4567-e89b-12d3-a456-426614174000
          format: uuid
          pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
          type: string
        processorId:
          example: 123e4567-e89b-12d3-a456-426614174000
          format: uuid
          pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
          type: string
        fileName:
          type: string
        contentType:
          example: application/pdf
          type: string
      required:
      - fileName
      - processorId
      - projectId
      type: object
    JobResponse:
      example:
        jobId: 123e4567-e89b-12d3-a456-426614174000
        createdAt: 2024-01-05T16:15:50Z
        completedAt: 2024-01-05T16:15:50Z
        processorId: 123e4567-e89b-12d3-a456-426614174000
        projectId: 123e4567-e89b-12d3-a456-426614174000
        status: pending
      properties:
        jobId:
          example: 123e4567-e89b-12d3-a456-426614174000
          format: uuid
          pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
          type: string
        status:
          enum:
          - pending
          - processing
          - completed
          - failed
          - cancelled
          type: string
        projectId:
          example: 123e4567-e89b-12d3-a456-426614174000
          format: uuid
          pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
          type: string
        processorId:
          example: 123e4567-e89b-12d3-a456-426614174000
          format: uuid
          pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
          type: string
        createdAt:
          example: 2024-01-05T16:15:50Z
          format: date-time
          type: string
        completedAt:
          example: 2024-01-05T16:15:50Z
          format: date-time
          type: string
      required:
      - createdAt
      - jobId
      - processorId
      - projectId
      - status
      type: object
    UploadUrlResponse:
      example:
        jobId: 123e4567-e89b-12d3-a456-426614174000
        uploadUrl: https://openapi-generator.tech
      properties:
        jobId:
          example: 123e4567-e89b-12d3-a456-426614174000
          format: uuid
          pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
          type: string
        uploadUrl:
          description: Pre-signed URL for file upload
          format: uri
          type: string
      required:
      - jobId
      - uploadUrl
      type: object
...
    IdResponse:
      example:
        id: 123e4567-e89b-12d3-a456-426614174000
      properties:
        id:
          example: 123e4567-e89b-12d3-a456-426614174000
          format: uuid
          pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
          type: string
      required:
      - id
      type: object
...
    ParseResponse:
      example:
        created_at: 2024-01-05T16:15:50Z
        parse_id: parse_id
      properties:
        parse_id:
          type: string
        created_at:
          example: 2024-01-05T16:15:50Z
          format: date-time
          type: string
      required:
      - created_at
      - parse_id
      type: object
...
    ReadResponse:
      example:
        created_at: 2024-01-05T16:15:50Z
        parse_id: parse_id
      properties:
        parse_id:
          type: string
        created_at:
          example: 2024-01-05T16:15:50Z
          format: date-time
          type: string
      required:
      - created_at
      - parse_id
      type: object
...
    _jobs_sync_post_request:
      properties:
        projectId:
          format: uuid
          type: string
        processorId:
          format: uuid
          type: string
        file:
          format: binary
          type: string
      required:
      - file
      - processorId
      - projectId
      type: object
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      description: JWT token authentication
      scheme: bearer
      type: http